pyproject.toml
[tool.poetry]
name = "memidebot"
version = "1.0.0"
description = ""
authors = ["Not4xp"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
python-telegram-bot = "^13.11"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"